requirements.txt
adxl345==1.0.4
asttokens==2.4.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorzero==2.0
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
fonttools==4.53.0
gpiozero==2.0.1
idna==3.7
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.5
lgpio==0.2.2.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pigpio==1.78
pillow==10.3.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
random2==1.0.2
requests==2.32.3
RPi.GPIO==0.7.1
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
smbus==1.1.post2
stack-data==0.6.3
tflite-runtime==2.14.0
traffic-light-classifier==1.0.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
